requirements.txt
twikit
google-generativeai # Or latest stable version
python-dotenv
aiofiles # For async file operations (optional but good practice)
